Fixed MediumDriverBase.Read and FloppyMediumDriver.cachedRead

+++ amiga_disk_devices/components/drivers/floppy_medium_driver.go
@@ -589,7 +589,7 @@ func (fmd *FloppyMediumDriver) cachedRead(floppyMedium *medium.FloppyMedium, pat
 	data, n, err := utils.FileUtilsInstance.FileReadBytes(
 		"",
 		ofst,
-		uint64(toReadSize),
+		toReadSize,
 		0,
 		0,
 		handle)
